load friendList
