+++ app/src/constant/HttpProperty.js
@@ -2,3 +2,7 @@ const BASE_URL = 'http://ec2-3-34-97-95.ap-northeast-2.compute.amazonaws.com';
 const VERSION = '/v1';
 
 export const SERVER_URL = BASE_URL + VERSION;
+
+export const GROUP_URL=SERVER_URL + '/group';
+export const SIGN_URL=SERVER_URL + '/sign';
+export const USER_URL=SERVER_URL + '/users';
